moved createConflict() to a separate file
