+++ test/integration/setup.js
@@ -9,7 +9,6 @@ const { noop } = require(appRoot + '/lib/util/util');
 const { task } = require(appRoot + '/lib/task/task');
 const authenticateUser = require('../util/authenticate-user');
 const testData = require('../data/xml');
-const { exhaust } = require(appRoot + '/lib/worker/worker');
 
 // knex things.
 const config = require('config');
@@ -220,36 +219,4 @@ const withClosedForm = (f) => async (service) => {
   return f(service);
 };
 
-const createConflict = async (user, container) => {
-  await user.post('/v1/projects/1/forms/simpleEntity/submissions')
-    .send(testData.instances.simpleEntity.one)
-    .set('Content-Type', 'application/xml')
-    .expect(200);
-
-  await exhaust(container);
-
-  await user.patch('/v1/projects/1/datasets/people/entities/12345678-1234-4123-8234-123456789abc?force=true')
-    .send({ data: { age: '99' } })
-    .expect(200);
-
-  await user.post('/v1/projects/1/forms?publish=true')
-    .send(testData.forms.updateEntity)
-    .set('Content-Type', 'application/xml')
-    .expect(200);
-
-  // all properties changed
-  await user.post('/v1/projects/1/forms/updateEntity/submissions')
-    .send(testData.instances.updateEntity.one)
-    .set('Content-Type', 'application/xml')
-    .expect(200);
-
-  await exhaust(container);
-};
-
-module.exports = {
-  testService, testServiceFullTrx,
-  testContainer, testContainerFullTrx,
-  testTask, withClosedForm,
-  createConflict
-};
-
+module.exports = { testService, testServiceFullTrx, testContainer, testContainerFullTrx, testTask, withClosedForm };
